Chore: Bump example to Flutter 3.29
