+++ example/android/app/src/main/kotlin/com/example/example/MainActivity.kt
@@ -2,4 +2,4 @@ package com.example.example
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity()
+class MainActivity : FlutterActivity()
